requirements.txt
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
Django==4.2.16
django-cors-headers==4.4.0
django-environ==0.11.2
django-js-asset==2.2.0
django-mptt==0.14.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
pkgutil_resolve_name==1.3.10
PyJWT==2.9.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
zipp==3.20.2
